Bug 37061: allow to configure how is data copied from host to component parts
This patch adds a new system preference to allow to configure how is data copied from host to component parts. The rules are defined in YAML format.
If left empty, the default behavior is to copy the host data to the component parts as hard coded in the code.

Test plan:
1) Apply the patch
2) Go to Administration -> Global system preferences
3) Search for PrepareHostField
4) Fill the rules in YAML format as in the example in the description
5) Save
6) Create a new child record
7) See that the data is copied from host to the component parts according to the rules