Bug 22322: Self check-in without log in
One should be able to return loans via SCO without
login required. This patch adds check in box where
patrons can check in their loans with barcode.

To test:
1. Find patron with loan or add one.
2. Go to /cgi-bin/koha/sco/sco-main.pl.
3. Return loan by passing items barcode.
=> Confirm loans is returned succesfully.
4. Attempt to return item again.
=> Confirm that "Item not checked in:..." message
is displayed.
5. Attempt to return empty barcode field.
=> Confirm that "Please enter barcode." message is
displayed.

Sponsored-by: Koha-Suomi Oy

Bug 22322: Don't log login failure to action_logs when page is refreshed

If sysprefs AuthFailureLog and SelfCheckoutByLogin
are enabled and one logs in to the SCO from OPAC
or refreshes the SCOs login page, this is added to the
action_logs table as an authentication failure.
We shouldn't log when this happens since these are
not real login attempts.

To test:
1. Make sure you have AuthFailureLog enabled.
2. Check how many rows are returned from database
with following query:
SELECT count(*) FROM action_logs WHERE module = "AUTH"
and action = "FAILURE";
3. Enable syspref SelfCheckoutByLogin.
4. Navigate to SCO page and login to OPAC when asked.
5. After SCO login page is displayed, check database again.
=> Result has now increased by 1.
6. Refresh page.
=> Result has again increased by 1.
7. Apply this patch, restart services if needed.
8. Login to SCO again from OPAC again or just refresh the page.
9. Check database.
=> Confirm that result hasn't increased.
10. Refresh page and check database.
=> Confirm that result hasn't increased.
11. Confirm that login to SCO still works as it should.

Sponsored-by: Koha-Suomi Oy

Bug 22322