Bug 41500: Do not drop value from Expired hold charge containing decimal with fraction when editing a rule
When one edits a circulation rule where Expired hold charge value
contains decimal with fractions other than 0 (e.g. 1.30), value is
not copied to editing row. This happens because column is considered
as column which cannot contain nothing else than integer value and
is therefore left blank if fraction is used.

To reproduce:
1. Add or modify existing circulation rule and set Expired hold charge value as 1.30.
2. Edit the rule you just created or modified.
=> Note that Expired hold charge is not displayed.
=> When saving the rule note that Expired hold charge is now overwritten with blank value.
3. Apply this patch.
4. Set Expired hold charge value as 1.30.
5. Edit the rule.
=> Note that Expired hold charge is now displayed correctly.

Sponsored-by: Koha-Suomi Oy